refactor(validation-icon): move to internal folder
BREAKING CHANGE: ValidationIcon has moved to src/__internal__/
