+++ docs/validations.stories.mdx
@@ -1,23 +1,22 @@
-<Meta
-  title="Documentation/Validations"
-/>
+<Meta title="Documentation/Validations" />
 
 # Getting started with validations, Formik and Yup
 
 ### Contents
-* [Carbon validation overview](#carbon-validation-overview)
-  * [Inputs](#inputs)
-  * [Grouped inputs](#grouped-inputs)
-  * [Validation icon placement](#validation-icon-placement)
-* [Formik overview](#formik-overview)
 
+- [Carbon validation overview](#carbon-validation-overview)
+  - [Inputs](#inputs)
+  - [Grouped inputs](#grouped-inputs)
+  - [Validation icon placement](#validation-icon-placement)
+- [Formik overview](#formik-overview)
 
 ## Carbon validation overview
+
 Carbon provides inputs which are able to visually indicate three states of validations out of the box.
 
-* **error**  –  meant to be a blocking validation, displays 2px wide red border and validation icon if message is provided.
-* **warning**  –  meant to be a non-blocking validation, displays 1px wide orange border and validation icon if message is provided.
-* **info**  –  meant to be a non-blocking validation, displays 1px wide blue border and validation icon if message is provided.
+- **error**  –  meant to be a blocking validation, displays 2px wide red border and validation icon if message is provided.
+- **warning**  –  meant to be a non-blocking validation, displays 1px wide orange border and validation icon if message is provided.
+- **info**  –  meant to be a non-blocking validation, displays 1px wide blue border and validation icon if message is provided.
 
 ### Inputs
 
@@ -35,10 +34,16 @@ Passing a `boolean` to these props will display a border and validation icon wit
 <Textbox name="name" value={value} onChange={onChange} error={true} />
 ```
 
-The only exception from this is the `DateRange` component which displays separate validation statuses on either `start` or `end` dates  -  to achieve this the `DateRange` component accepts `startError`, `endError`, `startWarning`, `endWarning`, `startInfo` and `endInfo` props. 
+The only exception from this is the `DateRange` component which displays separate validation statuses on either `start` or `end` dates - to achieve this the `DateRange` component accepts `startError`, `endError`, `startWarning`, `endWarning`, `startInfo` and `endInfo` props.
 
 ```js
-<DateRange name="name" value={value} onChange={onChange} startError="Start message" endError="End message" />
+<DateRange
+  name="name"
+  value={value}
+  onChange={onChange}
+  startError="Start message"
+  endError="End message"
+/>
 ```
 
 ### Grouped inputs
@@ -50,7 +55,13 @@ Passing validation props on a group component will display a validation icon on 
 The interface is the same - it's up to the consumer to decide whether to display validation on the group label/legend or on the individual component.
 
 ```js
-<RadioButtonGroup legend="Legend" name="name" value={value} onChange={onChange} error="Validation on group legend">
+<RadioButtonGroup
+  legend="Legend"
+  name="name"
+  value={value}
+  onChange={onChange}
+  error="Validation on group legend"
+>
   <RadioButton value="1" />
   <RadioButton value="2" />
 </RadioButtonGroup>
@@ -58,7 +69,7 @@ The interface is the same - it's up to the consumer to decide whether to display
 
 ```js
 <RadioButtonGroup name="name" value={value} onChange={onChange}>
-  <RadioButton value="1"  error="Validation on first grouped component" />
+  <RadioButton value="1" error="Validation on first grouped component" />
   <RadioButton value="2" />
 </RadioButtonGroup>
 ```
@@ -66,16 +77,18 @@ The interface is the same - it's up to the consumer to decide whether to display
 ### Validation icon placement
 
 Most of the input components support two variants of validation icon placement:
-* by the component or next to it - default behaviour
-* next to the component label or legend
+
+- by the component or next to it - default behaviour
+- next to the component label or legend
 
 To decide which placement is preferred `validationOnLabel` or `validationOnLegend` props can be used.
 
 ### Form component
 
 Carbon provides a Form component which augments an HTML `<form>` element with:
-* footer - a container for rendering "Save", "Cancel" and other custom buttons; it can also be made sticky if wanted
-* validation summary - graphical indicator of errors and warnings
+
+- footer - a container for rendering "Save", "Cancel" and other custom buttons; it can also be made sticky if wanted
+- validation summary - graphical indicator of errors and warnings
 
 For more information check [Form in Storybook](https://carbon.sage.com/?path=/docs/design-system-form)
 
@@ -86,7 +99,7 @@ For more information check [Form in Storybook](https://carbon.sage.com/?path=/do
 
 Formik provides a small set of React components and hooks, of which the key ones are:
 
-* [**Formik**](https://jaredpalmer.com/formik/docs/api/formik)  –  The central component of Formik, which wraps an
+- [**Formik**](https://jaredpalmer.com/formik/docs/api/formik)  –  The central component of Formik, which wraps an
   HTML form, and accepts props such as
   [initialValues](https://jaredpalmer.com/formik/docs/api/formik#initialvalues-values),
   [onSubmit()](https://jaredpalmer.com/formik/docs/api/formik#onsubmit-values-values-formikbag-formikbag-void),
@@ -100,11 +113,11 @@ Formik provides a small set of React components and hooks, of which the key ones
   [validateForm()](https://jaredpalmer.com/formik/docs/api/formik#validateform-values-any-promise-formikerrors-values),
   [validateField()](https://jaredpalmer.com/formik/docs/api/formik#validatefield-field-string-void) and
   [setValues()](https://jaredpalmer.com/formik/docs/api/formik#setvalues-fields-field-string-any-void)).
-* [**Form**](https://jaredpalmer.com/formik/docs/api/form),
+- [**Form**](https://jaredpalmer.com/formik/docs/api/form),
   [**Field**](https://jaredpalmer.com/formik/docs/api/field),
   [**ErrorMessage**](https://jaredpalmer.com/formik/docs/api/errormessage)  –  Helper components (powered by
   [React Context](https://reactjs.org/docs/context.html)) which reduce boilerplate code.
-* [**useField()**](https://jaredpalmer.com/formik/docs/api/useField)  –  A hook to help with building reusable input
+- [**useField()**](https://jaredpalmer.com/formik/docs/api/useField)  –  A hook to help with building reusable input
   primitive components.
 
 **Note:** Formik is only intended to be used with
@@ -113,7 +126,6 @@ controlled by supplying a `value` prop (or a `checked` prop for `radio` and `che
 use Formik with [uncontrolled components](https://reactjs.org/docs/uncontrolled-components.html) (by not supplying a
 controlling `value`/`checked` prop, and only supplying `name` `onChange` `onBlur` props), but this is strongly
 **discouraged** ⚠️
-
 
 ### Form validation using Formik
 
@@ -124,9 +136,7 @@ Formik supports three kinds of validation:
 2. [**Form-level: schema-based**](https://jaredpalmer.com/formik/docs/guides/validation#validationschema)  –  Performs
    validation using a [Yup](https://github.com/jquense/yup) schema.
 3. [**Field-level**](https://jaredpalmer.com/formik/docs/guides/validation#field-level-validation)  –  Separate
-   `validate()` functions are supplied for each form field.  This is the closest equivalent functionality of Carbon's
-   [`withValidation`](src/components/validations/with-validation.hoc.js) HOC which attaches validation functions to an
-   individual form field.
+   `validate()` functions are supplied for each form field.
 
 ## Code examples
