+++ src/__experimental__/components/input/input-presentation.d.ts
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { ValidationPropTypes } from "../../../components/validations";
+import { ValidationPropTypes } from "../../../__internal__/validations";
 
 export interface CommonInputPresentationProps extends ValidationPropTypes {
   /** If true, the component will be disabled */
