+++ src/__experimental__/components/radio-button/radio-button-group.d.ts
@@ -1,8 +1,10 @@
 import * as React from "react";
 import { MarginProps } from "styled-system";
-import { ValidationPropTypes } from "../../../components/validations";
+import { ValidationPropTypes } from "../../../__internal__/validations";
 
-export interface RadioButtonGroupProps extends ValidationPropTypes, MarginProps {
+export interface RadioButtonGroupProps
+  extends ValidationPropTypes,
+    MarginProps {
   /** Breakpoint for adaptive legend (inline labels change to top aligned). Enables the adaptive behaviour when set */
   adaptiveLegendBreakpoint?: number;
   /** Breakpoint for adaptive spacing (left margin changes to 0). Enables the adaptive behaviour when set */
