+++ src/__internal__/fieldset/fieldset.component.js
@@ -7,7 +7,7 @@ import {
   StyledFieldsetContent,
   StyledLegendContainer,
 } from "./fieldset.style";
-import ValidationIcon from "../../components/validations/validation-icon.component";
+import ValidationIcon from "../validations/validation-icon.component";
 import { InputGroupBehaviour, InputGroupContext } from "../input-behaviour";
 import { filterStyledSystemMarginProps } from "../../style/utils";
