+++ src/__internal__/validations/validation-icon.d.ts
