+++ src/__internal__/validations/validation-icon.style.js
@@ -1,7 +1,7 @@
 import styled from "styled-components";
 import { margin } from "styled-system";
 import BaseTheme from "../../style/themes/base";
-import StyledIcon from "../icon/icon.style";
+import StyledIcon from "../../components/icon/icon.style";
 
 const ValidationIconStyle = styled.div`
   background: none;
