+++ src/components/accordion/accordion.component.js
@@ -22,7 +22,7 @@ import {
   StyledAccordionContent,
 } from "./accordion.style";
 import Button from "../button";
-import ValidationIcon from "../validations";
+import ValidationIcon from "../../__internal__/validations";
 
 const Accordion = React.forwardRef(
   (
