+++ src/components/button-toggle-group/button-toggle-group.style.js
@@ -1,6 +1,6 @@
 import styled, { css } from "styled-components";
 import { StyledButtonToggleLabel } from "../button-toggle/button-toggle.style";
-import ValidationIconStyle from "../validations/validation-icon.style";
+import ValidationIconStyle from "../../__internal__/validations/validation-icon.style";
 import baseTheme from "../../style/themes/base";
 
 const ButtonToggleGroupStyle = styled.div`
