+++ src/components/checkbox/checkbox.style.js
@@ -9,7 +9,7 @@ import StyledCheckableInputSvgWrapper from "../../__internal__/checkable-input/c
 import StyledLabel, {
   StyledLabelContainer,
 } from "../../__internal__/label/label.style";
-import StyledValidationIcon from "../validations/validation-icon.style";
+import StyledValidationIcon from "../../__internal__/validations/validation-icon.style";
 import StyledFormField from "../../__experimental__/components/form-field/form-field.style";
 import StyledIcon from "../icon/icon.style";
 import baseTheme from "../../style/themes/base";
