+++ src/components/fieldset/fieldset.style.js
@@ -2,7 +2,7 @@ import styled from "styled-components";
 import { margin } from "styled-system";
 
 import FormFieldStyle from "../../__experimental__/components/form-field/form-field.style";
-import ValidationIconStyle from "../validations/validation-icon.style";
+import ValidationIconStyle from "../../__internal__/validations/validation-icon.style";
 import StyledIcon from "../icon/icon.style";
 import baseTheme from "../../style/themes/base";
 import CheckboxStyle from "../checkbox/checkbox.style";
