+++ src/components/numeral-date/numeral-date.d.ts
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { MarginProps } from "styled-system";
-import { ValidationPropTypes } from "../../../components/validations";
+import { ValidationPropTypes } from "../../__internal__/validations";
 
 interface DayMonthDate {
   dd: string;
