+++ src/components/switch/__internal__/switch-slider.style.js
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 
 import baseTheme from "../../../style/themes/base";
 import SwitchSliderPanel from "./switch-slider-panel.style";
-import StyledValidationIcon from "../../validations/validation-icon.style";
+import StyledValidationIcon from "../../../__internal__/validations/validation-icon.style";
 
 const StyledSwitchSlider = styled.span`
   ${({ checked, disabled, size, theme, error, warning, info }) => css`
