+++ src/components/text-editor/text-editor.spec.js
@@ -23,7 +23,7 @@ import Toolbar from "./__internal__/toolbar";
 import guid from "../../utils/helpers/guid/guid";
 import Label from "../../__internal__/label";
 import LabelWrapper from "./__internal__/label-wrapper";
-import ValidationIcon from "../validations";
+import ValidationIcon from "../../__internal__/validations";
 import { isSafari } from "../../utils/helpers/browser-type-check";
 
 jest.mock("../../utils/helpers/browser-type-check");
